backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==4.0.10
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
pillow==10.4.0
PyJWT==2.9.0
pymongo==3.12.2
pytz==2024.2
sqlparse==0.2.4
tzdata==2024.2

